% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot of accident coordinates for given state and year on state map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric value representing state in FARS dataset}

\item{year}{A numeric value of year}
}
\value{
This functions returns a map of the input state with accident
cooridnates plotted onto the map.
Latitude greator than 90 and longitude 900 is not plotted
If input state is not found in dataset, error is returned with message as
"invalid state number <state number>"
If no accidents are present in dataset for given state, a message is returned
as "no accidents to plot"
}
\description{
This function plots coordinates represented by latitude and longitude for
a given input state and year
}
\examples{
tryCatch({fars_map_state(1,2013)},error=function(e){return(NULL)})

}
