% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read and convert file to data frame}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name of the input csv file}
}
\value{
If input file doesnt exist, this function will stop execution
and return error else it will return tibble data frame object of input file
}
\description{
This function reads a csv file and converts to a tibble data frame
}
\examples{
tryCatch({fars_read("accident_2015.csv.bz2")}, error = function(e){NULL})

}
