% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Consolidate month & year data from multiple input files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector of year values for which data is required}
}
\value{
If input file doesnt exist for any corresponding input year a warning 
will be returned indicating that the input year is invalid. 
For all valid years, a list containing month and year fields of each input 
year is returned.
}
\description{
This function extracts and consolidates month and year fields from
multiple input files where each input file contains data for 1 year
}
\examples{
fars_read_years(c(2017,2018))
fars_read_years(2018)

}
