% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize month & year data from multiple input files}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector of year values for which data is required}
}
\value{
This functions returns a tibble data frame giving counts summarized
by month and year, where the intersection of a row representing month
and column representing year, gives summarized counts for that particular
month and year
}
\description{
This function summarizes consolidated month and year fields from
multiple input files where each input file contains data for 1 year
}
\examples{
fars_summarize_years(c(2013,2014))

}
