% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\docType{package}
\name{farsfunctions-package}
\alias{farsfunctions}
\alias{farsfunctions-package}
\title{Read, summarize and map US traffic accident (FARS) data.}
\description{
farsfunctions provides a set of functions to read US Department of Transportation
accident data from the Fatality Accident Reporting System (FARS).
}
\details{
Functions that might be of interest from \pkg{farsfunctions} include\cr
\code{\link{fars_read_years}} to access accident data\cr
\code{\link{fars_map_state}} to plot a map of accident locations.
Refer to the vignette for more information.
}
