% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate file name from input numeric value}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Numeric value of year}
}
\value{
This function returns a string value of file name
}
\description{
This function reads a numeric value and creates a file name string
}
\examples{
make_filename(2018)

}
