% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydatagr15.R
\docType{data}
\name{mydatagr15}
\alias{mydatagr15}
\title{Arabidopsis FARS data on accidents in 2015}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(mydatagr15)
}
\description{
Data from a FARS dataset in
50 variables and several thousand observations
}
\references{
Coursera datasets from the course Build R packages (available in August 2020)
#' @examples
str(mydatagr15)
}
\keyword{datasets}
