% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfuntions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{draw the map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the number of the state you want to draw}

\item{the}{year of the data you want to show}
}
\value{
a map with points
}
\description{
draw the map by year and state number
}
\details{
show a stop \code{"invalid STATE number: state.num"} if the state number is not is the data set. show a message \code{"no accident to plot"} if there is no data for the state in that year.otherwise, a map with points in \code{year}
}
\examples{
fars_map_state(1,2015)

}
