% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map FARS accidents by state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{FARS code of the state}

\item{year}{Year of interest}
}
\value{
A plot for the selected state and year showing the location of fatal accidents
}
\description{
This function provides a map for a scpecific state and year, that shows the location of
fatal accidents recorded in FARS.
}
\details{
If the state number is not corect, the function does not return the plot.
Longitudes>900 and latitudes>90 are considered as missing data.
}
\examples{
# Map accidents in NY (code 36) for 2015
fars_map_state(36, 2015)

}
