% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw Map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State, identified by its number.}

\item{year}{A year. The function will try to find the correct FARS file for the year
in the current directory.}
}
\value{
The function plots the map of accidents in a state from the FARS files for the given year.
}
\description{
Reads and plot FARS data for the given state and year.
}
\details{
The function check if the year data exists and the state (identified by its state number) exists in the year data.
Also sanitizes the location data (longitude and latitude). The function stops with error if the given state number
is incorrect (the state number is not in the FARS data files for the given year). A message is shown if no accidents
were registered for that year and state.

FARS files should be on the working directory.
}
\examples{
\dontrun{
fars_map_state( 25, 2014 )
}
}
