% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarsFunctions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{takes a state-number}

\item{year}{a year--not a list of years}
}
\value{
a map showing accident locations for the year
}
\description{
Creates a low-quality map of accidents in a state for a given year
}
\note{
ERRORS: Will give errors (and warnings) for invalid state or year inputs.
      Valid state inputs include most of the numbers from 1:56, excluding 3, 7, 14, 43, & 52.
      Many return bad maps. Valid year inputs include 2013, 2014, 2015.
      Will give a warning message and no map if there are no accidents to plot
}
