% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot a map of all fatal motor vehicle accidents for a specific year in a specified state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer of state ID number to plot a map for.}

\item{year}{Integer of the year to plot motor vehicle accidents from.}
}
\value{
A map of a state with points of GPS coordinates of motor vehicle accidents from a specific year.
}
\description{
makes a map of all fatal motor vehicle accidents for a specified year and specified state.
}
\details{
Throws an error if specified state ID numbers and years are invalid.
 Gives a warning if there are no fatal motor vehicle accidents in a given state and year.
}
\examples{
\dontrun{fars_map_state(15, 2013)}

}
\seealso{
\url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
