% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_hr.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Mapping FARS data for a specific state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state number}

\item{year}{The year for which the data is plotted}
}
\value{
A plot of FARS data on a map
}
\description{
The function shows FARS data on a map for a specific state and year.
If the state number is invalid, the error message is shown.
}
\examples{
\dontrun{
fars_map_state(50, 2013)
}

}
