% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map State Motor Vehicle Fatalities.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the number refering to the state of interest}

\item{year}{the year of interest - string or integer}
}
\value{
doesn't return anything but prints a map
}
\description{
Map State Motor Vehicle Fatalities.
}
\note{
state motor vehicle fatalities given a year and state id number.

won't work if referring to a year that has no data associated or a
state id without a relevant state associated
}
\examples{
library(mapdata)
fars_map_state(12, 2014)
fars_map_state(36, 2014)

}
