% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Show accidents on a state map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num, }{an integer, must be less than 57.}

\item{year, }{an integer or string coercible to an integer}
}
\value{
A map of the state, where accidents are marked
}
\description{
Given a year and a state number, this functions creates an image of the state
with the accidents marked. An error is given if either the state number or the
year is not valid. A message is provided if there were no accidents in a
paritcular state and year.
}
\examples{
\dontrun{
fars_map_state(40,2014)
}
}
