% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Display a map of accidents in a US state for a given year from FARS data}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{numeric value for the US state to be mapped - must be of length 1 or an error message is returned. If
the state value is not found in the accident file an error message is returned. For a list of numbers assigned
to the states see the FARS Analytical Users Manual.}

\item{year}{numeric value for the year to be mapped - must be of length 1 or an error message is returned}
}
\description{
\code{fars_map_state} reads FARS accident data for the year requested. If the file for that year is found,
the function plots a map of the state then plots the location of accidents in that year using the latitude and
longitude provided in the data file. If the accident file is not found an error message is returned.
}
\examples{
\dontrun{
fars_map_state(17, 2016) #map accidents in Illinois for the year 2016
}

}
\references{
for more information on FARS data see https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars
}
\seealso{
\code{\link{make_filename}} for the file nameing convention used\cr
\code{\link{fars_read}} for the data file input
}
