% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Locate the place of fatal injuries on map
This function displays the place of the fata linjuries by state and year on a map
an error message is given if the \code{state.num} is invalid}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer or character variable giving the state number}

\item{year}{integer or character variable giving the year}
}
\value{
None
}
\description{
Locate the place of fatal injuries on map
This function displays the place of the fata linjuries by state and year on a map
an error message is given if the \code{state.num} is invalid
}
\examples{
fars_map_state(1, 2015)
fars_map_state("1"," 2015")
}
\seealso{
\code{\link{fars_read}}

\code{\link{make_filename}}
}
