% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state function}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{is the US state number from where we will pick the information to be analized}

\item{year}{is the selected year of the data collected desired state number and year the selected data collected from the US National Highway Traffic Safety 
Administration's Fatality Analysis Reporting System, which is a nationwide census providing the American public 
yearly data regarding fatal injuries suffered in motor vehicle traffic crashes needs to be ploted.}
}
\value{
a plot of the accidents for a given state and year.
}
\description{
\code{fars_map_state(state.num,year)} returns a plot of the information loaded for a given state and year.
If the \code{state.num} is invalid this function stops execution and prints an error message.
 @importFrom dplyr filter
 @importFrom maps map
}
\examples{



\dontrun{
fars_map_state()
fars_map_state("10,2014")
}
}

