% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Visualize the location of accidents in given state for a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state numeric code}

\item{year}{The year of interest}
}
\value{
plot of the occurance of the accidents
}
\description{
Visualize the location of accidents in given state for a given year
}
\examples{
\dontrun{
fars_map_state(15, 2013)
}

}
