% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create a map of accidents in a state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A character string or numeric object specifying the number of the state
the function will use to filter out the data and subsequently draw a map}

\item{year}{A character string or numeric object giving the specific year value
the function will use to create a new file name, read a .csv data file and subsequently
create a map of traffic accidents locations}
}
\value{
This function returns a map of the specified state with
traffic accidents locations based on the specified year. If the state
number of invalid, the error message is printed out. If there is no
accidents to plot in a given year in a given state, the message is
printed out.
}
\description{
This is a function that draws a map of traffic accidents in a given year and state
based on the specified arguments. It first uses make_filename and fars_read functions
to create a .csv filename and subsequently read it. Then it filters out the data based
on the specified state. Finally, the function plots the traffic accident locations in
the given state using maps and graphics packages
}
\note{
.csv data file needs to be in your current directory
as the function does not take the whole file path but
filename only
}
\examples{
fars_map_state(1, 2013)
fars_map_state("1", 2013)
fars_map_state("1", "2013")
fars_map_state(1, "2013")

}
