% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Visualize accidents by state
This function plots accident locations as dots in a given US state for a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{This integer corresponds to a US state}

\item{year}{An integer correspondonds to year of interest. Function will return an error if more than one
year is input as an argument}
}
\value{
This function creates a plot showing locations of accidents in a given US state and year
}
\description{
Visualize accidents by state
This function plots accident locations as dots in a given US state for a given year
}
\examples{
fars_map_state(6,2013)
}
