% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot Accident Data}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric. The number of the state for accident data to be
plotted.}

\item{year}{A numeric. The year of the accident data to be plotted.}
}
\value{
This function returns a geographical plot of the specified state, defined
   by the \code{state.num}. Accidents are plotted as dots on the map according to
   data from the specified \code{year}. Each dot on the map represents one
   accident. If the specified \code{state.num} does not exist in the data, the
   function will stop executing and return an error message. If there are no
   accidents to plot for the specified \code{state.num} and \code{year}, this
   function will return a message informing the user of such a case.
}
\description{
This function plots a geographical map of the accidents that occurred in a
specified state and a specified year.
}
\examples{
\dontrun{fars_map_state(1, 2015)}

}
