% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot a map of accidents in a given state for a given year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a state in numeric format}

\item{year}{a year in numeric format or a list of such years}
}
\value{
this function returns a map on the graphical device, a message
"no accidents to plot", or a message saying that the provided state number is invalid.
}
\description{
This is a function that takes a state number and a year and plots
accidents on the states's map
}
\note{
requires the package maps to be loaded, otherwise produces an error.
For some states, the function fails with the message "nothing to draw: all regions out of bounds",
e.g. for state.num = 2.
}
\examples{
\dontrun{
fars_map_state(6, 2015) }

}
