% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_read}
\alias{fars_read}
\title{Read file with FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{\itemize{
\item name of the file (may be written with the path). The filename argument/input
for the function is a string with the name of the file to be opened
}}
}
\value{
A data frame with data readed from the csv file, or an error if the
file does not exists.

Function fars_read returns a tibble MxN where M - number of observations; N - number of columns
(50 for the accidents data files)
}
\description{
Open file related with the accident data files 2013-2015

Function fars_read can read .csv files and .csv archived in .bz2 format and other
formats compatible for the basic readr package for its function read_csv

Note that in the accidents data set only years 2013, 2014, and 2015 exist.
}
\details{
The functions work with the data taken from the .csv file (including .bz2 archived), stored on disk,
from the \strong{US National Highway Traffic Safety Administration's} (US NHTSA) \emph
{Fatality Analysis Reporting System} (FARS), the US nationwide census, containing the
US Public Annual Data, regarding fatal accidents occured in motorized
vehicle (like automobiles or bikes) traffic crashes.

For more information, see the source links (US NHTSA offial site and Wikipedia):
\itemize{
\item{\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}}
\item{\url{https://en.wikipedia.org/wiki/Fatality_Analysis_Reporting_System}}
}

In case if the file does not exist an error message is produced and execution stops.
}
\note{
To generate file name use: \code{\link{make_filename}}
}
\examples{
\dontrun{
library(dplyr)
library(readr)
setwd("C:/Users/Your_Path/")
data = fars_read("accident_2013.csv.bz2")
data = fars_read("C:/Users/Your_Path/accident_2013.csv.bz2")
head(data)
\dontrun}
}
\seealso{
\link{make_filename}
}
