% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfuntions.R
\name{fars_read}
\alias{fars_read}
\title{read the data}
\usage{
fars_read(filename)
}
\arguments{
\item{data}{filename you want to analyse}

\item{excited}{Logical value specifying whether the file exists}
}
\value{
a data frame named "data" or "file \code{data} does not exist"
}
\description{
This is a funtion to read the data you need and set it as a data frame
}
\examples{
\dontrun{
fars_read(system.file("data", "accident_2013.csv.bz2", package = "fars"))
accident_2014 <- fars_read(system.file("data", "accident_2014.csv.bz2", package = "fars"))
}

}
