% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the CSV file to read}
}
\value{
A tbl_df type of data containing the data from the CSV file
}
\description{
This function reads a CSV file containing Fatality Analysis Reporting System (FARS)
data and converts it to a dplyr tibble.
}
\examples{
# Read 2013 accident data
data_2013 <- fars_read("accident_2013.csv")

}
