% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read}
\alias{fars_read}
\title{Read FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the data file to be loaded}
}
\value{
This function returns a data frame table composed of the data from the
file specified by filename.
}
\description{
This is a function that takes a filename of a data file as an argument,
reads the data from the file, and shapes the data into a data frame table.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}

}
