% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarsFunctions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{of data to be read}
}
\value{
a data.frame of the file read
}
\description{
This function reads data, and returns it in a data.frame
}
\note{
ERRORS: If the file does not exist, it stops the program and sends an error message to that effect.
                  Works improperly and gives a warning if given a list of years
}
