% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in FARS data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string of the data file path.}
}
\value{
This function returns a data frame table of the FARS data.
}
\description{
Read in a Facility Analysis Reporting System data file.
}
\details{
Throws an error if file cannot be found.
}
\examples{
\dontrun{fars_read("accident_2015.csv.bz2")}

}
\seealso{
\url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
