% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_hr.R
\name{fars_read}
\alias{fars_read}
\title{Reading FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to the csv file as a character string.}
}
\value{
A tibble containing the data
}
\description{
The function reads FARS (Fatality Analysis Reporting System) data from a csv file.
If the file does not exist, an error message is shown.
}
\examples{
\dontrun{
fars_read("path/to/accident_2013.csv.bz2")
}

}
