% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS Data.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Character string giving the filename of the data.}
}
\value{
A tibble containing the FARS data.
}
\description{
Read FARS Data.
}
\note{
reads FARS data into the environment.

won't work if file name doesn't exist
}
\examples{
filename <- 'accident_2013.csv.bz2'
fars_read(filename)

}
