% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename, }{name of file containing data in csv format}
}
\value{
a dataframe of class tbl_df
}
\description{
Loads a datafile from disk. If no path is provided the current directory is
used. An error is given if the file does not exist.
}
\examples{
\dontrun{
fars_read("./data/accident_2013.csv.bz2")
}
}
