% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_3_building_packages_week2_functions.R,
%   R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\alias{fars_read}
\title{Reads data}
\usage{
fars_read(filename)

fars_read(filename)
}
\arguments{
\item{filename, }{name of file containing data in csv format}

\item{filename, }{name of file containing data in csv format}
}
\value{
a dataframe of class tbl_df

a dataframe of class tbl_df
}
\description{
Loads a datafile from disk. If no path is provided the current directory is
used. An error is given if the file does not exist.

Loads a datafile from disk. If no path is provided the current directory is 
used. An error is given if the file does not exist.
}
\examples{
fars_read("data/accident_2013.csv.bz2")
\donttest{fars_read("C:/path/farsfunctions/data/accident_2013.csv.bz2")}

fars_read("accident_2013.csv.bz2")
fars_read("C:/Users/My_filepath/accident_2013.csv.bz2")

}
