% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads the named csv file into a data table}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character name of the csv file to be read - must be a vector of length 1 or an error
message is returned.}
}
\value{
data frame coerced from the csv file input
}
\description{
\code{fars_read} is a simple function that reads a csv file and returns
a data table. If a file for a given year is not found a warning is issued
showing the year that is not valid.
}
\details{
This routine calls read_csv and returns a 'tbl_df' table data frame. The routine has
no expectations for (nor awareness of) the FARS data structures.
}
\examples{
\dontrun{
df <- fars_read("data/fars_data.csv")
my.data.frame <- fars_read("data/accident_2002.csv.bz2")
}

}
