% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read .csv file into a data frame}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of data file as a character string.}
}
\value{
The data file as a data frame.
}
\description{
The function loads US National Highway Traffic Safety Administration data and reads comma separated values file into a data frame.
If file doesn't exist, a warning message will appear.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}
}
