% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read the fars data into R
This function reads fars data stored as csv files into the working directory
and returns a tibble or dataframe
if the data do not exist an error message is given
You can customize the data to read with the \code{filename} argument}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a character string giving the filename to read}

\item{return}{returns the data as tibble/dataframe or an error message if the filename does not exist}
}
\description{
Read the fars data into R
This function reads fars data stored as csv files into the working directory
and returns a tibble or dataframe
if the data do not exist an error message is given
You can customize the data to read with the \code{filename} argument
}
\examples{
\dontrun{
fars_read("accident_2015.csv.bz2")
or make the filename with: \\code{\\link{make_filename}}
make_filename(2015) \%>\% fars_read()
}
}
\seealso{
\code{\link{make_filename}}
}
