% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in NHTSA FARS dataset CSV file as a data frame table (tbl_df) object}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string filename that corresponds to the CSV to be read in}
}
\value{
This function returns a tbl_df object
}
\description{
This is a function that takes a filename of a CSV as an argument and reads it in
as a data frame table. Successful execution will result in a data frame table
object containing the CSV data.
}
\details{
It will first check if the filename is valid and print a
warning and stop if it is not. The function will generate an error for invalid CSV
inputs.
}
\examples{
\dontrun{
fars_read('accident_2013.csv.bz2')
}
}
