% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_read}
\alias{fars_read}
\title{Victor Rainier Cruz Perez Roxygen2 Documentation exercise
FARS read function
********************** Instructions for testing the Package Functionality Accuracy ******************
Use  "test_dir("~/Desktop/CursoR2/farsfunctions/R/.")" to run all the test available in the fars_functions Package's R 
folder (test located inside the test_fars_functions.R script)
or
Use "test_file("~/Desktop/CursoR2/farsfunctions/R/test_fars_functions.R")" to run all the test available inside the
testing script test_fars_functions.R located inside the R folder of the fars_functions Package. The code "fars_functions.R"
must be sourced before executing the testing 
********************** Instructions for testing the Package Functionality Accuracy ******************}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{from where the data will be read.}
}
\value{
a table wich contains the data from the US National Highway Traffic Safety Administration's 
Fatality Analysis Reporting System, which is a nationwide census providing the American public yearly 
data regarding fatal injuries suffered in motor vehicle traffic crashes.
}
\description{
Use the Use_vignette once to generate the vignette folder in the package's skeleton where the vignette documentation
made with rmarkdown will be made
}
\details{
use_vignette("farsfunctionsdoc")
Use the use_testthat() once to generate the test folder in the package's skeleton where the test scripts will be placed

use_testthat()

\code{fars_read(filename)} returns a table with the information loaded from the filename provided as argument.
}
\examples{
fars_read("/Users/1/Desktop/data/fars_data.zip")


\dontrun{
fars_read()
}


}

