% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reading the Fars Data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{name of the file of the dataset}
}
\value{
dataframe converted from the read csv
}
\description{
reading the csv file and returning a dataframe
}
\examples{
\dontrun{
  fars_read("accident_2013.csv.bz2")
}
}
