% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Read .csv data files}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name
of a data file to read}
}
\value{
This function returns a data frame tibble which
is ready to parse
}
\description{
This is a simple function that reads a .csv data file
based on the supplied file name. Additionally,
it converts loaded data into data frame tibble.
It stops program execution and print an error message
if non-existing file name is supplied.
}
\note{
.csv data file needs to be in your current directory
as the function does not take the whole file path but
filename only
}
\examples{
fars_read("accident_2013.csv.bz2")

}
