% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read file
This is a function that reads data from a file and converts it to table data frame format. It returns error
if file does not exist}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string with the file name containing data to be read by the function}
}
\value{
This function returns a data frame containing the data in the file
}
\description{
Read file
This is a function that reads data from a file and converts it to table data frame format. It returns error
if file does not exist
}
\examples{
fars_read(accident_2013.csv.bz2)
}
