% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read CSV File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string. The name of the CSV file to be read.}
}
\value{
If the \code{filename} exists, this function will return a dataframe of the
   data contained in the CSV file. If the \code{filename} does not exist, the
   function will stop executing and return an error message.
}
\description{
This function first checks if the specified comma separated values filename exists,
reads it, and finally returns a data frame table.
}
\examples{
\dontrun{fars_read("accident_2015.csv.bz2")}

}
