% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a csv data file into a table dataframe (tbl_df)}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the dataset to read in}
}
\value{
This function returns a table dataframe (tbl_df) if it is given
   a valid name or a notification or a notification "file 'filename' does not exist"
}
\description{
This is a function that reads a csv data file into a table dataframe (tbl_df).
You are supposed to provide the name of a dataset (csv file) to read as the argument.
}
\note{
if there is no file wit hthe given filename, the function will return an error
}
\examples{
\dontrun{
filname("somedata.csv") }

}
