% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector with a list of years}
}
\value{
A data.frame including entries in data by month,
or NULL if the \code{year} is not valid
}
\description{
Ancillary function used by \code{fars_summarize_years}
Primary function to the
}
\examples{
\dontrun{
setwd("C:/Users/Your_Path/")
fars_read_years(2013)
\dontrun}
}
\seealso{
\link{fars_read}

\link{make_filename}

\link{fars_summarize_years}
}
