% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfuntions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{read data for years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{datas of years that your want to read, returned by \code{\link{make_filename}}}

\item{excites}{the Logical value specifying whether the file of the year exists}
}
\value{
a data named "dat" or "NULL" and "invalid year: \code{years}" if the \code{years} does not exist
}
\description{
read the data for specific years and save the data to "dat"
}
