% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years}
}
\value{
The information corresponding to each month of the selected years
}
\description{
This function reads FARS data files for multiple years and selects
the MONTH and year columns from each file.
}
\details{
The function returns NULL for invalid years
}
\examples{
# Read data for multiple years
fars_13_14_15 <- fars_read_years(c(2013, 2014, 2015))
fars_13_9999 <- fars_read_years(c(2013, 9999))

}
