% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarsFunctions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a list of years to be access--should be a numeric.  Can also work with a single string or numeric}
}
\value{
A list of data.frames containing accident data
}
\description{
Takes a list of years, and returns a list containing data.frames of accidents
}
\note{
ERRORS: Will give a warnings for any invalid years and return null values (valid years are 2013, 2014 and 2015)
}
