% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_hr.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reading FARS data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years, years as integers.}
}
\value{
A list of tibbles, each containing data for a specific year
}
\description{
The function reads FARS data for multiple years and combines them into a tibble.
If the year is not valid, the warning message is shown.
}
\examples{
\dontrun{
fars_read_years(c(2013, 2014))
fars_read_years(2013:2015)
}

}
