% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads a FARS file for a given vector of years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector of years corresponding to FARS file by year.
It also accepts a charcter vector and converts to numeric.}
}
\value{
It returns a list of data frame with two columns and where each list
is for the input years. The two columns are \code{MONTH} and \code{year}.
If the file for the given year is not found, it returns a list of
\code{NULL} with a warning message of invalid year:  \code{years}.
}
\description{
Reads a FARS file for a given vector of years.
}
\details{
\code{fars_read_years} function imports the FRAS file for a given
vector years, and it uses \code{\link{make_filename}} to create the
correct file name associated with \code{years}.And then imports the file using
\code{fars_read} function, and then adds(mutates) a varaible \code{year}. Then selects and displays
two variables \code{MONTH}, \code{year}.
}
\examples{
fars_read_years(2015)

}
