% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS files for one or more years.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a string or integer showing which years of data to load}
}
\value{
Returns a list of tibbles. Each tibble contains the year and month
rom the observations
}
\description{
Read FARS files for one or more years.
}
\note{
produces a list of tibbles of FARS data, given an input vector of years.

won't work if referring to a year that has no data associated
}
\examples{
fars_read_years(years = c(2013, 2014, 2015))

}
