% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_3_building_packages_week2_functions.R,
%   R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\alias{fars_read_years}
\title{Read in data for multiple years}
\usage{
fars_read_years(years)

fars_read_years(years)
}
\arguments{
\item{years, }{a list of years (should be a list of integers)}

\item{years, }{a list of years (should be a list of integers)}
}
\value{
A list of lists of data for each valid year in the input list 'years'. The data

A list of lists of data for each valid year in the input list 'years'. The data
}
\description{
Given years as input, this function generates proper filenames using
make_filename(), then reads the files using fars_read(). The data is subsequently
manipulated, by creating an extra column, year, and selecting the colunms MONTH and year.
A warning message is provided if data does not exist for a given year. This
function will never fail, as whole function is run with tryCatch.

Given years as input, this function generates proper filenames using 
make_filename(), then reads the files using fars_read(). The data is subsequently 
manipulated, by creating an extra column, year, and selecting the colunms MONTH and year. 
A warning message is provided if data does not exist for a given year. This 
function will never fail, as whole function is run with tryCatch.
}
\examples{
fars_read_years(c(2013,2014))
fars_read_years(c(2013,2014))
my_yrs <- c(2013,2014,2015))
fars__read_years(my_yrs)

}
