% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Get year and month columns from one or more FARS accident data files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector of one or more years of files to read}
}
\value{
a list of one or more data frames (one for each year). Each data frame has two columns from each row of the files found:
\itemize{\item MONTH is the numeric month (1 through 12)
\item year is the numeric year}
}
\description{
\code{fars_read_years} reads one or more years of FARS accident data to provide
the year and MONTH columns. If a file is not found a warning is returned that the
year is not valid (if the file does not contain a column named MONTH and a column named
year the same warning is issued).
}
\examples{
\dontrun{
my_df <- fars_read_years(2014:2016)
}

}
\references{
for more information on FARS data see https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars
}
\seealso{
\code{\link{make_filename}} for the file nameing convention used
}
