% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Return the month and year corresponding to each FARS observation}
\usage{
fars_read_years(years)
}
\arguments{
\item{year}{A year passed as an argument in the form of an integer or string for which 
the data should be read in}
}
\value{
This function returns a list
}
\description{
This is a function that takes a year in the form of an integer or string as an argument 
and then uses that value to read in a CSV as a tbl_df for data of that given year and 
and returns the month and year corresponding to each observation.
}
\details{
This function will return an error message for invalid years.
}
\examples{
fars_read_years(2013)
fars_read_years('2013')

}
