% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years function}
\usage{
fars_read_years(years)
}
\arguments{
\item{...}{desired list of years for selecting data collected from the US National Highway Traffic Safety 
Administration's Fatality Analysis Reporting System, which is a nationwide census providing the American public 
yearly data regarding fatal injuries suffered in motor vehicle traffic crashes.}
}
\value{
a table wich contains a slide of the data from the US National Highway Traffic Safety Administration's 
Fatality Analysis Reporting System for a desired years
}
\description{
\code{fars_read_years(years)} returns a table with the information loaded for a list of years.
    If an invalid year is passaed as \code{years} parameter, the
    data frame returned will be NULL an a warning will be delivered
}
\examples{
fars_read_years(c(2013,2014)


\dontrun{
fars_read_years()
fars_read_years("2013","2014")
}
}

