% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS data from multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a string vector of years}
}
\value{
A dataframe of the fars data in month and year, if the invalid year in
entered then NULL would be returned
}
\description{
if the years provided are valid, then the function would load the years
}
\examples{
\dontrun{
fars_read_years(c(2012,2013))
}

}
