% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read specific .csv data files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of character strings or numeric objects
giving the specific year value the function will use to create
a new file name and subsequently read a .csv data file}
}
\value{
This function returns a list of manipulated data frame
tibbles that result from data parsing within tryCatch block.
If the supplied argument is incorrect, the error block is executed
and NULL value fills up the appropriate list slot.
}
\description{
This is a function that reads a .csv data file based on
the specifically supplied argument. It uses lapply function
to call make_filename function and subsequently intends to
read a .csv data file using fars_read function based on the
newly created file name. The function uses tryCatch block
to evaluate the validity of the supplied argument. Within
the block, the function manipulates the data frame tibble
using dplyr package.
}
\note{
.csv data file needs to be in your current directory
as the function does not take the whole file path but
filename only
}
\examples{
fars_read_years("2013")
fars_read_years(c(2013, 2014))

}
