% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Obtain data from years
This function returns the month number and year from the file corresponding to the variable 'years'.
If a year is not included in the file, an error is returned}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{An integer of list thereof containing at least one year of interest.}
}
\value{
This function returns the data corresponding to the years in years
}
\description{
Obtain data from years
This function returns the month number and year from the file corresponding to the variable 'years'.
If a year is not included in the file, an error is returned
}
\examples{
fars_read_years(2013)
fars_read_years(c(2010,2011))
}
