% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read Accident Data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector. The years of accident data to be read.}
}
\value{
This function returns a list of data frames, with each data frame
   corresponding to each of the \code{years} specified. The data frames only
   contain information about the month and year of each accident that occured. If
   the file for any of the specified \code{years} cannot be read, that index will
   be returned as null along with an appropriate warning message.
}
\description{
This function reads accident data for each of the years specified. It then selects
just the month and year of each accident and returns a list of data frames for
each corresponding year.
}
\examples{
\dontrun{fars_read_years(c(2013, 2014, 2015))}

}
