% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in the data on MONTH variable from multiple FARS datafiles for different years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a year in numeric format or a list of such years}
}
\value{
This function returns a list, each element of which is a two-column table (tibble)
with one column describing the year and the other column describing the month for each observation
in the FARS dataset.
}
\description{
This is a function that takes a year or a list of years in numeric format, reads in
the FARS data files for the corresponding years, and returns a list of tables for each
year. Each table contains only the data for two variables: the year and months for which
FARS data are available in that year
}
\note{
providing multiple years that are not in the list format as an argument will cause an error.
Also, the function will produce an error if the dplyr package is not loaded.
}
\examples{
\dontrun{
fars_read_years(2015)
fars_read_years(list(2014, 2015)) }


}
