% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data by years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector with a list of years to summarize by.}
}
\value{
A data.frame with number of accidents by years summarized by month
}
\description{
This simple function summarizes yearly accidents data, by month
}
\examples{
\dontrun{
setwd("C:/Users/Your_Path/")
plot(fars_summarize_years(2013))
fars_summarize_years(c(2013, 2015))
\dontrun}
}
\seealso{
\link{fars_read_years}
}
