% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfuntions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{summarize the data of the year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{the data of the year you want to summarize,returned by \code{\link{fars_read_years}}}
}
\value{
a data named "dat_list" and a spread sheet
}
\description{
summarize number of the data of the year by year and month and make the spread sheet
}
\examples{
fars_summarize_years(2014,2015)

\dontrun{
fars_summarize_years(2013)
fars_summarize_years(2013:2015)
}

}
