% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data by years and months}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years}
}
\value{
A tibble where each row corresponds to a month (from 1=january to 12=december)
and each column is a year. It contains the number of accidents per month/year.
}
\description{
This function creates a summary of the number of accidents by month and year
}
\examples{
fars_summarize_years(c(2013, 2014, 2015))

}
