% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarsFunctions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{takes a year or list of years to be accessed--can be as numerics or strings}
}
\value{
a data.frame giving the number of accidents by year and month
}
\description{
Returns a dataframe with the number of accidents by year and month
}
\note{
ERRORS: Will give errors (and warnings) for invalid year inputs (valid years include 2013, 2014, and 2015)
}
