% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count number of motor vehicle accidents by year for selected years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Numeric vector of years to summarize.}
}
\value{
Returns a tbl_df of the number of motor vehicle accidents for each selected year.
}
\description{
Counts the number of fatal motor vehicle accidents for each month of the specified years in the FARS data set.
}
\details{
Throws an error if specified years are invalid.
}
\examples{
\dontrun{fars_summarize_years(c(2013,2014,2015))}

}
\seealso{
\url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
