% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_hr.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizing FARS data for multiple years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years, years as integers.}
}
\value{
A tibble summarizing the number of accidents per month for each year
}
\description{
The function summarizes FARS data for multiple years, counting accidents per month.
}
\examples{
\dontrun{
fars_summarize_years(c(2013, 2014))
fars_summarize_years(2013:2015)
}

}
