% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Produce a Summary of FARS Files.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Va string or integer showing which years of data to summarize}
}
\value{
a tibble with an id column as monht, and the remaining columns for
selected yars. Values in year columns are number of observation that month /
year
}
\description{
Produce a Summary of FARS Files.
}
\note{
produces a summary tibble of FARS years and months given a vector of years.

won't work if referring to a year that has no data associated
}
\examples{
fars_summarize_years(years = c(2013, 2014, 2015))
fars_summarize_years(years = 2015)
}
