% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes the FARS data by year and month.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Integer of years or acharacter coreciable to numeric.}
}
\value{
Returns a summary of accidents in a data frame where the first column
reprsents months. Each column are the input years and is filled by
the number of accidents happend in each montht. If the input year is
not in the data set it returns an error.
}
\description{
The \code{fars_summarize_years} function summarizes the number of accidents for
each year by month. The first column represents \code{MONTH} and the
other column are the input years. If the input year is not present in the data
it returns an error.
}
\examples{
fars_summarize_years(c(2013,2014,2015))
}
