% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_3_building_packages_week2_functions.R,
%   R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize data}
\usage{
fars_summarize_years(years)

fars_summarize_years(years)
}
\arguments{
\item{years, }{a list of years (should be a list of integers)}

\item{years, }{a list of years (should be a list of integers)}
}
\value{
A dataframe, with summary data of accidents for each month

A dataframe, with summary data of accidents for each month
}
\description{
Given a list of years, loads data associated with the years, selects and
summarizes relevant data.

Given a list of years, loads data associated with the years, selects and 
summarizes relevant data.
}
\examples{
fars_summarize_years(c(2014,2015))
fars_summarize_years(c(2014,2015))
}
