% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count the number of accidents in FARS data over the years requested}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector of one or more years of files to read}
}
\value{
a data frame with a column for the 12 months followed by a column for each year returned. The rows
are the numeric month followed by a count of the accidents for that year and month so the table has at most 12 rows.
For example, if years 2015 #' and 2016 are requested and there are 2371 accidents in 2015 and 2344 accidents in 2016,
the data frame would have column names of 'MONTH', '2015' and '2016' and start something like this:
\tabular{lllll}{
MONTH \tab 2015 \tab 2016\cr
1 \tab 2371 \tab 2344\cr
2 \tab 1983 \tab 2421\cr
}
with additional rows for the remaining months
}
\description{
\code{fars_summarize_years} reads one or more years of FARS accident data then summarizes a count
by month of the number of accidents. If a file is not found a warning is returned that the
year is not valid (if the file does not contain a column named MONTH and a column named
year the same warning is issued).
}
\examples{
\dontrun{
my_df <- fars_summarize_years(2015:2016)
}

}
\references{
for more information on FARS data see https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars
}
\seealso{
\code{\link{make_filename}} for the file nameing convention used\cr
\code{\link{fars_read_years}} for the file input
}
