% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Calculate number of accidents per month for the particular years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{desired list of years for selecting data collected from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System, which is a nationwide census providing the American public
yearly data regarding fatal injuries suffered in motor vehicle traffic crashes to be summarized.}
}
\value{
a summary table wich contains the number of accidents grouped by year and month of the given list of years
}
\description{
The function summarises  the accidents for the selected years and counts the number
of reports per month.
}
\examples{
\dontrun{
fars_summarize_years(2013:2015)
fars_summarize_years(c(2013, 2014, 2015))
fars_summarize_years(list(2013, 2014, 2015))
}

}
