% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Return a monthly summary of incidents}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A year passed as an argument in the form of an integer or string for which
the data should be read in.}
}
\value{
This function returns a tbl_df object
}
\description{
This is a function that takes a year in the form of an integer or string as an argument
and then uses that value to read in a CSV as a tbl_df for data of that given year and
and returns a monthly summary of the observations.
}
\details{
This function will return an error message for invalid years.
}
\examples{
\dontrun {
fars_summarize_years(2013)
fars_summarize_years('2013')
}

}
