% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years function}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{...}{desired list of years for selecting data collected from the US National Highway Traffic Safety 
Administration's Fatality Analysis Reporting System, which is a nationwide census providing the American public 
yearly data regarding fatal injuries suffered in motor vehicle traffic crashes to be summarized.}
}
\value{
a summary table wich contains the number of accidents grouped by year and month of the given list of years
}
\description{
\code{fars_summarize_years(years)} returns a summary table from the information loaded for a given list of years.
Because the list of years is limited; invalid years are not included in the data frame.
}
\examples{
fars_read_years(c(2013,2014)


\dontrun{
fars_read_years()
fars_read_years("2013","2014")
}
}

