% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizing the farsdata by months}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{the values of years to be summarizied}
}
\value{
a dataframe with monthly number of accidents and years
}
\description{
Monthly number of accident in each months in the years mentioned
}
\examples{
\dontrun{
fars_summarize_years(c(2012,2013))
}

}
