% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize traffic accident data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of character strings or numeric objects
giving the specific year value the function will use to call
fars_read_years function}
}
\value{
This function returns a tibble which results
from data manipulation using "tidyverse" packages. It groups
the data by year and month and summarizes it resulting in a number
of traffic accidents per month in the previously specified year
}
\description{
This is a function that summarizes the number of traffic accidents
based on the supplied argument. It takes advantage of
fars_read_years function. The list output of that function is then
parsed using "tidyverse" set of packages.
}
\note{
.csv data file needs to be in your current directory
as the function does not take the whole file path but
filename only
}
\examples{
fars_summarize_years("2013")
fars_summarize_years(c(2013, 2014))

}
