% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Obtain summary values of data
This function returns the number of accidents of a specific year or set of years for each month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{An integer of list thereof containing at least one year of interest.}
}
\value{
a data frame is returned containing the number of accidents per month (rows) and year (cols)
}
\description{
Obtain summary values of data
This function returns the number of accidents of a specific year or set of years for each month
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(c(2013,2014))
}
