% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize Accident Data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector. The years of accident data to be summarized.}
}
\value{
This function returns a single data frame. Each year from the \code{years}
   specified is allocated a separate column and each month is allocated a separate
   row, and each data point represents the total number of accidents that occured
   in that corresponding month and year.
}
\description{
This function summarizes accident data for the specified years. It returns a data
frame table with information about the number of accidents that occurred in each
month of each year.
}
\examples{
\dontrun{fars_summarize_years(c(2013, 2014, 2015))}

}
