% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize month data for multiple years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a year in numeric format or a list of such years}
}
\value{
this function returns a table (tibble) with one column for each given year and
up to twelve rows each representing a month.
}
\description{
This is a function that takes a year or a list of years in numeric format and
generates a summary table that describes the number of observations in FARS datafiles
for each month in each year.
}
\note{
if the list that is provided as an argument contains a year for which the dataset
is unavailable, it will produce an error.
}
\examples{
\dontrun{
fars_summarize_years(2015)
fars_summarize_years(list(2013, 2014, 2015)) }

}
