\name{farsfunctionspackage}
\alias{farsfunctionspackage}
\alias{farsfunctions}
\docType{package}
\title{
\packageTitle{farsfunctions}
}
\description{
\packageDescription{farsfunctions}
}
\details{

The DESCRIPTION file:This is a dummy test for making R packages
\packageDESCRIPTION{farsfunctions}
\packageIndices{farsfunctions}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{farsfunctions}

Maintainer: \packageMaintainer{farsfunctions}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ There are no other links to other man pages. ~~
}
\examples{

make_filename("2014")
fars_read_years(c(2013,2014))
fars_map_state(10,2014)
}
