\name{farsfunctionspackage}
\alias{farsfunctionspackage}
\alias{farsfunctions}
\docType{package}
\title{
\packageTitle{farsfunctions}
}
\description{
\packageDescription{farsfunctions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{farsfunctions}
\packageIndices{farsfunctions}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{farsfunctions}

Maintainer: \packageMaintainer{farsfunctions}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
