% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfuntions.R
\name{make_filename}
\alias{make_filename}
\title{mark the file by years}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year that you want to mark out}
}
\value{
the file named by the year you specified
}
\description{
mark the file by different years
}
\examples{
\dontrun{
make_filename(2013)
}

}
