% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate FARS filenames}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric value representing the year}
}
\value{
A filename "accident_xxxx.csv.bz2", where xxxx represents the year
}
\description{
This function generates a filename for FARS data based on the year (2013, 2014 or 2015)
}
\examples{
make_filename(2013)
make_filename(2014)
make_filename(2015)

}
