% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarsFunctions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year or a list of years to be access--either as strings or numerics.  Can also take lists of these.}
}
\value{
a string or list of strings indicating the filenames to be returned
}
\description{
creates a filename to access data accident data for specific years from the
             US National Highway Traffic Safety Administration's Fatality Analysis Reporting System
}
\note{
ERRORS: None, but will return useless filenames for any inputs except 2013, 2014, and 2015
}
