% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a file name for data in a specific year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Character String of the year that the data file is being made for.}
}
\value{
Returns a character string of the filename for the specified year.
}
\description{
Creates a filename to be read in with specified FARS data.
}
\examples{
\dontrun{make_filename(2015)}

}
\seealso{
\url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
