% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_hr.R
\name{make_filename}
\alias{make_filename}
\title{Creating file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year for which the filename is generated}
}
\value{
A character string representing the generated filename
}
\description{
The function generates the filename for a FARS data file, based on the input year.
}
\examples{
\dontrun{
make_filename(2013)
}

}
