% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{make_filename}
\alias{make_filename}
\title{Make FARS Filename.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year of interest - string or integer}
}
\value{
returns a string that is the proper filename for the FARS data for
that year
}
\description{
Make FARS Filename.
}
\note{
makes a properly formatted FARS filename given a year as input

won't work if year is not in integer or a string format
}
\examples{
make_filename(year = '2013')
make_filename(year = 2013)

}
