% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{make_filename}
\alias{make_filename}
\title{Makes a file name for a given year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Can be a numeric or chracter vector of years. if it is chracter, chnages to numeric using \code{as.integer} function.}
}
\value{
It returns a character of file name, like "accident_2013.csv.bz2".
}
\description{
\code{make_filename} function takes a year argument and creates a file name of the given year.
}
\examples{
\dontrun{
make_filename(2013)
}
}
