% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Construct filename}
\usage{
make_filename(year)
}
\arguments{
\item{year, }{should be an integer or a string coercible to an integer}
}
\value{
a string
}
\description{
Constructs a filename of the type accident_'year'_csv.z2, where year is
given as input.
}
\examples{
\dontrun{
make_filename(1996)
}
}
