% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_3_building_packages_week2_functions.R,
%   R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\alias{make_filename}
\title{Construct filename}
\usage{
make_filename(year)

make_filename(year)
}
\arguments{
\item{year, }{should be an integer or a string coercible to an integer}

\item{year, }{should be an integer or a string coercible to an integer}
}
\value{
a string

a string
}
\description{
Constructs a filename of the type accident_'year'_csv.z2, where year is
given as input.

Constructs a filename of the type accident_'year'_csv.z2, where year is 
given as input.
}
\examples{
make_filename(1996)

make_filename(1996)

}
