% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates a name for the accident csv file using a year value}
\usage{
make_filename(year)
}
\arguments{
\item{year}{one or more year values to be included in the file name string}
}
\value{
character vector of file names constructed from the years provided
}
\description{
\code{make filename} uses the input year to create a file name for the accident csv file. For example
if the year parameter has a value of 2012 the file name constructed will be "accident_2002.csv.bz2".
If the year parameter value cannot be coerced into a numeric integer value the file name will be
""accident_NA.csv.bz2"
}
\examples{
make_filename(2015)

}
