% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create filename for a compressed bz2 file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year of the accidents as an integer.}
}
\value{
The file name as a character string.
}
\description{
The function creates a file name using a year as an input.This function makes reading files
faster for the analysis of FARS data.
}
\examples{
\dontrun{
make_filename(year = 2013)
}

}
