% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make the filename of the fars data
This function takes the (\code{year} as a string or integer as input and
returns the filename of the fars data stored as csv files
The return value of this function can be used as input for the \code{fars_read function}}
\usage{
make_filename(year)
}
\arguments{
\item{year}{integer or string giving the year of the data}
}
\value{
a string giving the filename of the fars data
}
\description{
Make the filename of the fars data
This function takes the (\code{year} as a string or integer as input and
returns the filename of the fars data stored as csv files
The return value of this function can be used as input for the \code{fars_read function}
}
\examples{
make_filename(2015)
make_filename("2015")
}
