% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename function}
\usage{
make_filename(year)
}
\arguments{
\item{year}{desired year for having a slide of the data collected from the US National Highway Traffic Safety 
Administration's Fatality Analysis Reporting System, which is a nationwide census providing the American public 
yearly data regarding fatal injuries suffered in motor vehicle traffic crashes.}
}
\value{
a file name of a table wich contains a slide of the data from the US National Highway Traffic Safety Administration's 
Fatality Analysis Reporting System from a desired year. The file name is composed by a definend string and the given year.
}
\description{
\code{make_filename(year)} returns a table with the information loaded for a desired year.
}
\examples{


\dontrun{
make_filename()
}
}

