% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a file for the dataset}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year to be added to the filename}
}
\value{
a filename with the year entered
}
\description{
creates a name for the accident file based on the year provided.
}
\examples{
\dontrun{
make_filename(2018)
}

}
