% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Print the name of a file for a specific year
This functions prints the name of a file containing data from a specific year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer with the year of interest}
}
\value{
This function returns and prints the name of a file containing data from a specific year
}
\description{
Print the name of a file for a specific year
This functions prints the name of a file containing data from a specific year
}
\examples{
make_filename(2010)
}
