% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a filename in the format used for FARS data files}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A year in the numeric format}
}
\value{
This function returns a character string that corresponds to the name of
the FARS datafile for the given year
}
\description{
This is a function that takes a year in numeric format as an argument
and returns a filename of the corresponding FARS data file for the given year.
}
\note{
spaces, commas or other characters in the numeric argumet will retur an error
}
\examples{
make_filename(2015)

}
